requirements.txt
fastapi[standard]
psycopg2
SQLAlchemy
passlib[bcrypt]
argon2-cffi
pyjwt
pydantic-settings
alembic